backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
selenium==4.24.0
webdriver-manager==4.0.2
pandas==2.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.8.2